Create /root/.profile
This .profile script checks for running Rastari (i.e. is script run.sh running) and doesn't try to start another Rastari session. This allows to log in as root. In addition to this change, a user pi (UID 1000, password "raspberry") and group pi (GID 1000) should be added and the sudo command should be installed ("apt-get install sudo"), so one can login as user pi and switch to root by typing "sudo su -".